ExtractAudioFeatures, with_delta as float32 fix

+++ GeneratingDataset.py
@@ -759,7 +759,8 @@ class ExtractAudioFeatures:
 
     if self.with_delta:
       import librosa
-      deltas = [librosa.feature.delta(feature_data, order=i, axis=0) for i in range(1, self.with_delta + 1)]
+      deltas = [librosa.feature.delta(feature_data, order=i, axis=0).astype("float32")
+                for i in range(1, self.with_delta + 1)]
       feature_data = numpy.concatenate([feature_data] + deltas, axis=1)
       assert feature_data.shape[1] == self.get_feature_dimension()
